+++ internal/handlers/mediastream.go
@@ -65,6 +65,9 @@ func (h *Handler) HandleSaveMediastreamSettings(c echo.Context) error {
 //	@returns mediastream.MediaContainer
 //	@route /api/v1/mediastream/request [POST]
 func (h *Handler) HandleRequestMediastreamMediaContainer(c echo.Context) error {
+	if err := h.guardStrictLocalOnlyAction(c); err != nil {
+		return err
+	}
 
 	type body struct {
 		Path             string                 `json:"path"`             // The path of the file.
@@ -116,6 +119,9 @@ func (h *Handler) HandleRequestMediastreamMediaContainer(c echo.Context) error {
 //	@returns bool
 //	@route /api/v1/mediastream/preload [POST]
 func (h *Handler) HandlePreloadMediastreamMediaContainer(c echo.Context) error {
+	if err := h.guardStrictLocalOnlyAction(c); err != nil {
+		return err
+	}
 
 	type body struct {
 		Path             string                 `json:"path"`             // The path of the file.
@@ -154,6 +160,10 @@ func (h *Handler) HandlePreloadMediastreamMediaContainer(c echo.Context) error {
 }
 
 func (h *Handler) HandleMediastreamGetSubtitles(c echo.Context) error {
+	if err := h.guardStrictLocalOnlyAction(c); err != nil {
+		return err
+	}
+
 	c.Response().Header().Set("Cache-Control", "no-cache, no-store, must-revalidate")
 	c.Response().Header().Set("Pragma", "no-cache")
 	c.Response().Header().Set("Expires", "0")
@@ -161,6 +171,10 @@ func (h *Handler) HandleMediastreamGetSubtitles(c echo.Context) error {
 }
 
 func (h *Handler) HandleMediastreamGetAttachments(c echo.Context) error {
+	if err := h.guardStrictLocalOnlyAction(c); err != nil {
+		return err
+	}
+
 	// tell the client not to cache the response
 	c.Response().Header().Set("Cache-Control", "no-cache, no-store, must-revalidate")
 	c.Response().Header().Set("Pragma", "no-cache")
@@ -173,6 +187,10 @@ func (h *Handler) HandleMediastreamGetAttachments(c echo.Context) error {
 //
 
 func (h *Handler) HandleMediastreamDirectPlay(c echo.Context) error {
+	if err := h.guardStrictLocalOnlyAction(c); err != nil {
+		return err
+	}
+
 	client := "1"
 	return h.App.MediastreamRepository.ServeEchoDirectPlay(c, client)
 }
@@ -182,6 +200,10 @@ func (h *Handler) HandleMediastreamDirectPlay(c echo.Context) error {
 //
 
 func (h *Handler) HandleMediastreamTranscode(c echo.Context) error {
+	if err := h.guardStrictLocalOnlyAction(c); err != nil {
+		return err
+	}
+
 	client := "1"
 	return h.App.MediastreamRepository.ServeEchoTranscodeStream(c, client)
 }
@@ -205,6 +227,10 @@ func (h *Handler) HandleMediastreamShutdownTranscodeStream(c echo.Context) error
 //
 
 func (h *Handler) HandleMediastreamFile(c echo.Context) error {
+	if err := h.guardStrictLocalOnlyAction(c); err != nil {
+		return err
+	}
+
 	client := "1"
 	fp := c.QueryParam("path")
 	libraryPaths := h.App.Settings.GetLibrary().GetLibraryPaths()